OBDX-7 refactoring as Justin comments
